Adding Rabbit MQ for communication

+++ projects/src/main/java/WrokStream360/project/controller/ProjectController.java
@@ -1,17 +1,17 @@
 package WrokStream360.project.controller;
 
-import WrokStream360.project.modal.ProjectRequest;
-import WrokStream360.project.modal.ProjectResourceRequest;
-import WrokStream360.project.modal.ProjectResponse;
+import WrokStream360.project.modal.*;
 import WrokStream360.project.service.ProjectService;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.repository.query.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.List;
 
 @RestController
 @RequestMapping("/project")
@@ -21,7 +21,7 @@ public class ProjectController {
     private ProjectService projectService;
 
     // Create project in DB
-    @PostMapping("/add-project")
+    @PostMapping("/create-project")
     public ResponseEntity<Long> addProject(@RequestBody ProjectRequest projectRequest){
         long projectId = projectService.addProject(projectRequest);
       return new ResponseEntity<>(projectId , HttpStatus.CREATED);
@@ -45,13 +45,41 @@ public class ProjectController {
 
     @DeleteMapping("/{projectId}")
     public ResponseEntity<String> deleteProject(@PathVariable long projectId) {
-        // Call the deleteProjectById() method of the project service
+        // Check if the project exists
+        if (!projectService.exists(projectId)) {
+            return new ResponseEntity<>("Project not found", HttpStatus.NOT_FOUND);
+        }
+        // Delete the project and its resources
         projectService.deleteProjectById(projectId);
-
-        // Return a success response
-        return ResponseEntity.ok("Project and associated resources deleted");
+        return new ResponseEntity<>("Project and resources deleted successfully", HttpStatus.OK);
     }
 
+
+
+    @PutMapping("/{id}")
+    public ResponseEntity<String> updateProject(@PathVariable("id") Long projectId, @RequestBody ProjectRequest projectRequest) {
+        try {
+            projectService.updateProject(projectId, projectRequest);
+            return ResponseEntity.ok("Project updated successfully");
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to update project");
+        }
+    }
+
+
+    @PatchMapping("/update-budget-data/{id}")
+    public ResponseEntity<Long> updateBudgetData(@PathVariable("id") Long projectId, @RequestBody ProjectBudgetChange projectBudgetChange) {
+        long updateProjectBudgetData = projectService.updateBudgetData(projectId, projectBudgetChange);
+        return new ResponseEntity<>(updateProjectBudgetData, HttpStatus.OK);
+    }
+
+
+    @GetMapping
+    public List<AllProjectResponse> getAllProjects() {
+        return projectService.getAllProjects();
+    }
 
 
 }
